fully move config migration logic into the contribution
