+++ src/vs/platform/prompts/common/config.ts
@@ -123,58 +123,59 @@ export namespace PromptsConfig {
 		return [];
 	};
 
-	/**
-	 * TODO: @legomushroom
-	 */
-	// TODO: @legomushroom - merge object with potentially existent `promptFilesLocations` config
-	export const migrateOldSetting = (
-		configService: IConfigurationService,
-	) => {
-		const value = configService.getValue('chat.promptFiles');
-		// const value2 = read('promptFiles');
+	// /**
+	//  * TODO: @legomushroom
+	//  */
+	// // TODO: @legomushroom - remove?
+	// // TODO: @legomushroom - merge object with potentially existent `promptFilesLocations` config
+	// export const migrateOldSetting = (
+	// 	configService: IConfigurationService,
+	// ) => {
+	// 	const value = configService.getValue('chat.promptFiles');
+	// 	// const value2 = read('promptFiles');
 
-		if ((value === undefined) || (value === null)) {
-			return;
-		}
+	// 	if ((value === undefined) || (value === null)) {
+	// 		return;
+	// 	}
 
-		if (Array.isArray(value)) {
-			configService.updateValue('chat.promptFiles', true);
+	// 	if (Array.isArray(value)) {
+	// 		configService.updateValue('chat.promptFiles', true);
 
-			const locationsValue: Record<string, boolean> = {};
-			for (const filePath of value) {
-				if (typeof filePath !== 'string') {
-					continue;
-				}
-				const trimmedValue = filePath.trim();
-				if (!trimmedValue) {
-					continue;
-				}
+	// 		const locationsValue: Record<string, boolean> = {};
+	// 		for (const filePath of value) {
+	// 			if (typeof filePath !== 'string') {
+	// 				continue;
+	// 			}
+	// 			const trimmedValue = filePath.trim();
+	// 			if (!trimmedValue) {
+	// 				continue;
+	// 			}
 
-				locationsValue[trimmedValue] = true;
-			}
+	// 			locationsValue[trimmedValue] = true;
+	// 		}
 
-			configService.updateValue('chat.promptFilesLocations', locationsValue);
-			return;
-		}
+	// 		configService.updateValue('chat.promptFilesLocations', locationsValue);
+	// 		return;
+	// 	}
 
-		if (typeof value === 'object') {
-			configService.updateValue('chat.promptFiles', true);
-			configService.updateValue('chat.promptFilesLocations', value);
+	// 	if (typeof value === 'object') {
+	// 		configService.updateValue('chat.promptFiles', true);
+	// 		configService.updateValue('chat.promptFilesLocations', value);
 
-			return;
-		}
+	// 		return;
+	// 	}
 
-		if (typeof value === 'string') {
-			const booleanValue = asBoolean(value);
-			if (booleanValue !== undefined) {
-				return;
-			}
+	// 	if (typeof value === 'string') {
+	// 		const booleanValue = asBoolean(value);
+	// 		if (booleanValue !== undefined) {
+	// 			return;
+	// 		}
 
-			configService.updateValue('chat.promptFiles', true);
-			configService.updateValue('chat.promptFilesLocations', { [value]: true });
-			return;
-		}
-	};
+	// 		configService.updateValue('chat.promptFiles', true);
+	// 		configService.updateValue('chat.promptFilesLocations', { [value]: true });
+	// 		return;
+	// 	}
+	// };
 }
 
 /**
@@ -185,7 +186,7 @@ export namespace PromptsConfig {
  * 			be clearly mapped to a boolean (e.g., `"true"`, `"TRUE"`, `"FaLSe"`, etc.),
  * 			`undefined` for rest of the values
  */
-const asBoolean = (value: any): boolean | undefined => {
+export const asBoolean = (value: any): boolean | undefined => {
 	if (typeof value === 'boolean') {
 		return value;
 	}
